requirements.txt
#tensorflow==2.9
scikit-learn==1.6.1
PyYAML==6.0.2
flask==3.1.0
pandas==2.2.3
opencv-python-headless
ipykernel 
python-box
dill-0.3.9
opencv-python-headless==4.11.0.86
pip install pillow
-e .